refactor: update cosxuxi crawler imports
